fix: improve stereo shader with parallel cameras and auto-scaling
- Use parallel cameras instead of toe-in convergence (reduces eye strain)
- Add horizontal image shift (HIT) for convergence control via dof parameter
- Scale stereo base automatically when zoomed in to keep deviation bounded
- Compute camera distance from z-translation (independent of rotation)
- Enable key repeat for adjustment commands (eof/dof/gamma/contrast)
- Fix spdlog enum formatting warning
- Update defaults: eof=0.0075, dof=0.01 (conservative comfortable values)

+++ src/GobanModel.cpp
@@ -177,7 +177,7 @@ void GobanModel::onGameMove(const Move& move, const std::string& comment) {
         state.reason = move == Move::RESIGN ? GameState::RESIGNATION : GameState::DOUBLE_PASS;
         board.toggleTerritoryAuto(true);
         isGameOver = true;
-        spdlog::debug("Main Over! Reason {}", state.reason);
+        spdlog::debug("Main Over! Reason {}", static_cast<int>(state.reason));
     }
     else if (move == Move::PASS) {
         state.msg = state.colorToMove == Color::BLACK
